Broken out api config into own file
